create a file for CubeWriter avoid having twice the unit displayed
