+++ ticoi_env.yml
@@ -8,7 +8,6 @@ dependencies:
   - fiona=1.9
   - gdal
   - geopandas>=0.10.0
-  - glaft
   - h5netcdf
   - h5py>=3.11
   - jdcal
